refactor: move cdp.rs to tools/repl
